[#17] 좋아요 기능을 레디스로 캐시

+++ src/main/java/com/stoury/dto/FeedResponse.java
@@ -16,9 +16,9 @@ public record FeedResponse(Long feedId,
                            List<String> tagNames,
                            Double latitude,
                            Double longitude,
-                           int likes,
+                           long likes,
                            LocalDateTime createdAt) {
-    public static FeedResponse from(Feed uploadedFeed, int feedLikes) {
+    public static FeedResponse from(Feed uploadedFeed, long feedLikes) {
         return new FeedResponse(
                 uploadedFeed.getId(),
                 MemberResponse.from(uploadedFeed.getMember()),
